:seedling: Remove APIReader from  scope
Removing the APIReader from  scope, as it is not used in many places.
The occurences where it is used, we can replace it. Instead of using the
elaborate AcquireSecret function, we can just use the normal client.Get
method. We don't need to put finalizers or anything on these secrets
that we handle there, as they are managed by ClusterAPI anyway. Only the
HivelocitySecret has to be managed by us.