:seedling: Update associate device logic and include chooseDevice
We update the logic of the associate function to include chooseDevice
function. This is necessary, because we have otherwise two different
devices in the main Reconcile function, one which is associated, and one
which is chosen to be. This is confusing and is implemented better in
this way, when a fully associated device object is created within
associate and returned by that function.