:sparkles: Add DeviceTag object
Adding DeviceTag and DeviceTagKey objects that make it easier to handle
Hivelocity device tags. They are strings but we expect a certain
pattern. If this pattern is not matched, a tag is invalid. The new
objects have methods to check whether a tag string is valid or not. They
are easier to use than the previous utility functions.

On top, some small bugs have been fixed in the provisioning process..