:bug: Add failure domain and update shutdown logic
FailureDomains are required from CAPI but not yet set in the
HivelocityCluster object.

The shutdown process has to be adapted because of flaws in the
Hivelocity API. There is no reliable way of finding out whether a
machine is switched on or off or boots/shuts down. We maybe have to
continue to work on this.