:sparkles: Implementing state machine for deleting
We have the following states: When a machine is supposed to be deleted,
we first check in which provisioning state we are. When the provisioning
of the device is not yet finished, we can just remove the relevant tags
of the device. If it is provisioned already, then we need to shut the
device down and, after this is done, re-provision it with a default
image. In this way, the device is removed from the cluster.